Created abstract class for games with child classes for sp, mp and minichess
