+++ model/constants.js
@@ -32,8 +32,6 @@ module.exports = {
 	},
 
 	winCondition: {
-	    CHIPS: 0,
-	    FIGURES: 1,
 	    PAWN_TO_LAST_ROW: 2
 	    // FIG_TO_CENTER: 3
 	},
